+++ python/sglang/srt/layers/moe/topk.py
@@ -57,9 +57,6 @@ if _use_aiter:
 if _is_npu:
     import torch_npu
 
-if _is_npu:
-    import torch_npu
-
 
 class TopKOutput(NamedTuple):
     topk_weights: torch.Tensor
@@ -617,24 +614,6 @@ def select_experts(
     num_token_non_padded: Optional[torch.Tensor] = None,
     expert_location_dispatch_info: Optional[ExpertLocationDispatchInfo] = None,
 ) -> TopKOutput:
-    global_num_experts = router_logits.shape[-1]
-    # NOTE: now npu_moe_gating_top_k can only support `group_count=256` pattern,
-    # and generalization to more scenarios will be supported in the future.
-    if _is_npu and global_num_experts == 256:
-        topk_weights, topk_ids, _ = torch_npu.npu_moe_gating_top_k(
-            router_logits,
-            k=top_k,
-            bias=correction_bias.to(torch.float32),
-            k_group=topk_group,
-            group_count=num_expert_group,
-            group_select_mode=1,
-            renorm=0,
-            norm_type=1,
-            routed_scaling_factor=1,
-            eps=float(1e-20),
-        )
-        return topk_weights, topk_ids
-
     router_logits, correction_bias = (
         expert_location_dispatch.transform_select_experts_inputs(
             router_logits=router_logits,
