+++ python/sglang/srt/managers/schedule_batch.py
@@ -106,7 +106,6 @@ GLOBAL_SERVER_ARGS_KEYS = [
     "num_reserved_decode_tokens",
     "weight_loader_disable_mmap",
     "enable_triton_kernel_moe",
-    "quantization",
 ]
 
 # Put some global args for easy access
